feature(FrondEnd): Routes final Definition And Atomic Design implementation
