+++ src/Components/Atoms/Styles/BackGroundAuthComponentAtom.tsx
@@ -1,0 +1,12 @@
+export default function BackGroundAuthComponentAtom() {
+    return (
+        <>
+            <div
+            className="absolute top-0 w-full h-full bg-Gray-800 bg-no-repeat bg-full"
+            style={{
+              backgroundImage: "url('/img/register_bg_2.png')",
+            }}
+            ></div>
+        </>
+    )
+}
