+++ src/Components/Atoms/Titles/CredentialsMessage.tsx
@@ -1,0 +1,9 @@
+export default function CredentialsMessageComponentAtom() {
+    return (
+        <>
+            <div className="text-gray-400 text-center mb-3 font-bold">
+                <small>Or sign up with credentials</small>
+            </div>
+        </>
+    )
+}
