+++ src/Components/Atoms/Titles/SignInMessageComponentAtom.tsx
@@ -1,0 +1,11 @@
+export default function SignInMessageComponentAtom() {
+    return (
+        <>
+            <div className="text-center mb-3">
+                  <h6 className="text-gray-500 text-sm font-bold">
+                    Sign up with
+                  </h6>
+            </div>
+        </>
+    )
+}
