+++ src/Components/Footers/FooterAuth.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React from "react";
-
+import LinksComponentMolecule from "../Molecules/FooterAuth/LinksFooterComponentMolecule";
 
 export default function FooterAuth() {
   return (
@@ -10,55 +10,7 @@ export default function FooterAuth() {
       >
         <div className="container mx-auto px-4 py-4">
           <hr className="mb-6 border-b-1 border-gray-600" />
-          <div className="flex flex-wrap items-center md:justify-between justify-center">
-            <div className="w-full md:w-4/12 px-4">
-              <div className="text-sm text-gray-500 font-semibold py-1 text-center md:text-left">
-                Copyright © {new Date().getFullYear()}{" "}
-                <a
-                  href="https://github.com/CQuiroga"
-                  className="text-white hover:text-gray-300 text-sm font-semibold py-1"
-                >
-                  Camilo Quiroga
-                </a>
-              </div>
-            </div>
-            <div className="w-full md:w-8/12 px-4">
-              <ul className="flex flex-wrap list-none md:justify-end justify-center">
-                <li>
-                  <a
-                    href="#"
-                    className="text-white hover:text-gray-300 text-sm font-semibold block py-1 px-3"
-                  >
-                    Link 1
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="text-white hover:text-gray-300 text-sm font-semibold block py-1 px-3"
-                  >
-                    Link 2
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="text-white hover:text-gray-300 text-sm font-semibold block py-1 px-3"
-                  >
-                    Link 3
-                  </a>
-                </li>
-                <li>
-                  <a
-                    href="#"
-                    className="text-white hover:text-gray-300 text-sm font-semibold block py-1 px-3"
-                  >
-                    Link 4
-                  </a>
-                </li>
-              </ul>
-            </div>
-          </div>
+          <LinksComponentMolecule />
         </div>
       </footer>
     </>
