+++ src/app/Scores/Admin/[scoreId]/page.tsx
@@ -1,0 +1,14 @@
+
+'use client';
+
+import React from 'react';
+
+export default function ScoreAdminById() {
+  return (
+    <>
+        <p>Esto es ScoreAdmin By Id</p>
+        <p>Patch, get, put</p>
+       
+    </>
+  );
+}
