+++ src/app/Scores/Admin/page.tsx
@@ -1,0 +1,15 @@
+
+'use client';
+
+import React from 'react';
+
+export default function ScoreAdmin() {
+  return (
+    <>
+        <p>Esto es ScoreAdmin </p>
+        <p>Get de todos los scores</p>
+
+       
+    </>
+  );
+}
