+++ src/app/Scores/Profile/[id]/page.tsx
@@ -1,0 +1,14 @@
+
+'use client';
+
+import React from 'react';
+
+export default function ScoresProfileById({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+        <p>Esto es scores profile ById</p>
+
+
+    </>
+  );
+}
