+++ src/app/Scores/[scoreId]/page.tsx
@@ -1,0 +1,15 @@
+
+'use client';
+
+import React from 'react';
+
+export default function ScoreById() {
+  return (
+    <>
+        <p>Esto es scores by id</p>
+        <p>Puy y patch</p>
+
+
+    </>
+  );
+}
