+++ src/app/Users/Admin/[userId]/page.tsx
@@ -1,0 +1,21 @@
+
+'use client';
+
+import React from 'react';
+
+export default function UsersAdminById() {
+  return (
+    <>
+        <p>Esto es users Admin by id</p>
+        <p>patch</p>
+
+        <p>y put</p>
+
+        <p>Esto es users by id</p>
+
+        <p>Esto es users by id</p>
+
+
+    </>
+  );
+}
