+++ src/app/Users/Profile/[userId]/page.tsx
@@ -1,0 +1,21 @@
+
+'use client';
+
+import React from 'react';
+
+export default function UsersProfileById() {
+  return (
+    <>
+        <p>Esto es users by id</p>
+        <p>Esto es users by id</p>
+
+        <p>Esto es users by id</p>
+
+        <p>Esto es users by id</p>
+
+        <p>Esto es users by id</p>
+
+
+    </>
+  );
+}
