--- src/app/Users/Scores/page.tsx
@@ -1,11 +1,0 @@
-"use client";
-import Leaderboard from "game-score-frontend/Components/Pages/LeaderBoard";
-
-export default function LeaderboardPage () {
-
-    return (
-        <Leaderboard />
-    );
-};
-
-;
