+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from "next";
 import localFont from "next/font/local";
 import "./globals.css";
+import React from 'react';
 
 const geistSans = localFont({
   src: "./fonts/GeistVF.woff",
