+++ src/app/page.tsx
@@ -6,7 +6,5 @@ export default function Home() {
     <>  
     <Login />
     </>
-    
-   
   );
 }
